Added pre-created sqlite database to handle story data.
Removed XML files and all code associated with XML parsing. No longer applicable to the scope of this project.

Renamed classes and layouts to more meaningful, descriptive names.

Added a StoryView class to replace a LinearLayout for the main menu items. Now referred to as My Library and My Library items.

 Created a base adapter for all lists to implement. Changed all lists to RecyclerViews.

 Created and amended the three styles into something more unique. Styles now inherit from the new Material design theme.

 Long click on a My Library item takes you to its About screen.